refactor: main to split it into sub part and add time display on ssd1306

+++ src/time/sync_time.rs
@@ -14,7 +14,8 @@ pub fn sync_ntp() -> Result<()> {
 
 pub fn get_local_time(offset_secs: i32) -> (u8, u8, u8) {
     let timestamp_utc = SystemTime::now().duration_since(UNIX_EPOCH).unwrap().as_secs();
-    let local_secs = timestamp_utc + offset_secs as u64;
+    let local_secs = (timestamp_utc as i64 + offset_secs as i64) as u64;
+
 
     let seconds_per_day = local_secs % 86400;
     let hours = seconds_per_day / 3600;
